+++ example/lib/constants.dart
@@ -33,18 +33,19 @@ class Constants {
   static String elephantDreamStreamUrl =
       'https://devstreaming-cdn.apple.com/videos/streaming/examples/img_bipbop_adv_example_fmp4/master.m3u8';
   static String tokenEncodedHlsUrl =
-      'https://demo.unified-streaming.com/k8s/features/stable/video/tears-of-steel/tears-of-steel.ism/.m3u8';
+      'https://media.axprod.net/TestVectors/v7-MultiDRM-SingleKey/Manifest_1080p.mpd';
   static String tokenEncodedHlsToken =
-      'Bearer=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1cm46bWljcm9zb2Z0OmF6dXJlOm1lZGlhc2VydmljZXM6Y29udGVudGtleWlkZW50aWZpZXIiOiI5ZGRhMGJjYy01NmZiLTQxNDMtOWQzMi0zYWI5Y2M2ZWE4MGIiLCJpc3MiOiJodHRwOi8vdGVzdGFjcy5jb20vIiwiYXVkIjoidXJuOnRlc3QiLCJleHAiOjE3MTA4MDczODl9.lJXm5hmkp5ArRIAHqVJGefW2bcTzd91iZphoKDwa6w8';
+      'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ2ZXJzaW9uIjoxLCJjb21fa2V5X2lkIjoiYjMzNjRlYjUtNTFmNi00YWUzLThjOTgtMzNjZWQ1ZTMxYzc4IiwibWVzc2FnZSI6eyJ0eXBlIjoiZW50aXRsZW1lbnRfbWVzc2FnZSIsImZpcnN0X3BsYXlfZXhwaXJhdGlvbiI6NjAsInBsYXlyZWFkeSI6eyJyZWFsX3RpbWVfZXhwaXJhdGlvbiI6dHJ1ZX0sImtleXMiOlt7ImlkIjoiOWViNDA1MGQtZTQ0Yi00ODAyLTkzMmUtMjdkNzUwODNlMjY2IiwiZW5jcnlwdGVkX2tleSI6ImxLM09qSExZVzI0Y3Iya3RSNzRmbnc9PSJ9XX19.FAbIiPxX8BHi9RwfzD7Yn-wugU19ghrkBFKsaCPrZmU';
   static String widevineVideoUrl =
-      'https://storage.googleapis.com/wvmedia/cenc/h264/tears/tears_sd.mpd';
+      'https://storage.googleapis.com/shaka-demo-assets/sintel-widevine/dash.mpd';
   static String widevineLicenseUrl =
-      'https://proxy.uat.widevine.com/proxy?provider=widevine_test';
+      'https://cwip-shaka-proxy.appspot.com/no_auth';
   static String fairplayHlsUrl =
-      'https://fps.ezdrm.com/demo/hls/BigBuckBunny_320x180.m3u8';
+      'https://livesim2.dashif.org/livesim2/drm_EZDRM-1-key-cbcs/testpic_2s/Manifest.m3u8';
   static String fairplayCertificateUrl =
-      'https://github.com/jhomlala/betterplayer/raw/master/example/assets/eleisure.cer';
-  static String fairplayLicenseUrl = 'https://fps.ezdrm.com/api/licenses/';
+      'https://na-fps.ezdrm.com/demo/video/eleisure.cer';
+  static String fairplayLicenseUrl =
+      'https://fps.ezdrm.com/api/licenses/auth?px=A552CE&assetID=01234567-89AB-CDEF-0123-456789ABCDEF';
   static String catImageUrl = 'https://picsum.photos/id/237/1000/1000';
   static String dashStreamUrl =
       'https://bitmovin-a.akamaihd.net/content/MI201109210084_1/mpds/f08e80da-bf1d-4e3d-8899-f0f6155f6efa.mpd';